fix : 훅으로 처리했던것 revert(#95)
